Refactor GraphQL модуля.
