+++ modules/graphql/README.md
@@ -2,26 +2,18 @@
 
 ``` bash
 .
-├── GraphqlModule.php
-├── README.md
+├── base # Базовые класы мутаций и запросов
 ├── controllers
-│   └── GraphqlController.php
-└── schema # Схемы для GraphQL
-    ├── common # Общие схемы для mutation и query
-    │   ├── Response.php # Универсальная схема ответа для мутаций
-    │   ├── Types.php # Все наши кастомные типы
-    │   └── ValidationErrorType.php # Схема для ошибок валидации
+├── data # Типы данных которые мы создали Enum, Errors, Mutation, Query, Response ...
+└── schema # GraphQL схемы
+    ├── common # Базовые схемы для мутаций и запросов
     ├── mutations # Мутации
-    │   ├── BaseMutationType.php # Абстрактная мутация 
-    │   ├── MutationTrait.php # Расщирение для мутаций
-    │   ├── MutationType.php # Базовая GraphQL схема для мутаций
-    │   └── extended # Все наши кастомные схемы мутаций, по сути Models
-    │       └── ...
-    └── types # Запросы
-        ├── QueryType.php # Базовая GraphQL схема для запросов
-        ├── TypeTrait.php # Расширение для запросов
-        └── extended # Все наши кастомные схемы запросов, по сути Models
-            └── ...
+    │   ├── MutationType.php # Базовая схема
+    │   └── extended # Все наши схемы для мутаций моделей, партнёры, продукты ...
+    └── query # Запросы на получение данных
+        ├── QueryType.php # Базовая схема запросов
+        └── extended Все наши схемы для получения данных моделей, партнёры, продукты ...
+            └── enum # Отдельно Выделил схемы для перечислений, так как они не модели
 ```
 
 ### queries
