+++ modules/graphql/schema/mutations/extended/PartnerMutationType.php
@@ -4,23 +4,23 @@ declare(strict_types = 1);
 namespace app\modules\graphql\schema\mutations\extended;
 
 use app\models\partners\Partners;
-use app\modules\graphql\schema\mutations\MutationTrait;
-use app\modules\graphql\schema\mutations\BaseMutationType;
+use app\modules\graphql\base\BaseMutationType;
+use app\modules\graphql\data\ErrorTypes;
+use app\modules\graphql\data\MutationTypes;
 use GraphQL\Type\Definition\Type;
-use app\modules\graphql\schema\common\Types;
+use app\modules\graphql\data\QueryTypes;
+use yii\db\ActiveRecord;
 
 /**
  * Class PartnerMutationType
- * @package app\modules\graphql\schema\mutations
+ * @package app\modules\graphql\schema\mutations\extended
  */
 final class PartnerMutationType extends BaseMutationType
 {
-	use MutationTrait;
-
 	/**
 	 * {@inheritdoc}
 	 */
-	public ?string $model = Partners::class;
+	protected ?ActiveRecord $model;
 
 	/**
 	 * {@inheritdoc}
@@ -28,26 +28,13 @@ final class PartnerMutationType extends BaseMutationType
 	public const MESSAGES = ['Ошибка сохранения партнера', 'Партнер успешно сохранен'];
 
 	/**
-	 * {@inheritdoc}
+	 * PartnerMutationType constructor.
+	 * @param Partners $model
 	 */
-	public function __construct()
+	public function __construct(Partners $model)
 	{
-		parent::__construct([
-			'fields' => [
-				'update' => [
-					'type' => Types::validationErrorsUnionType(Types::partner()),
-					'description' => 'Обновление партнера',
-					'args' => $this->getArgs(),
-					'resolve' => fn(array $rootArgs, array $args = []): array => $this->update($rootArgs, $args),
-				],
-				'create' => [
-					'type' => Types::validationErrorsUnionType(Types::partner()),
-					'description' => 'Создание партнера',
-					'args' => $this->getArgs(),
-					'resolve' => fn(array $rootArgs, array $args = []): array => $this->create($args),
-				],
-			]
-		]);
+		$this->model = $model;
+		parent::__construct($this->getConfig());
 	}
 
 	/**
@@ -56,7 +43,7 @@ final class PartnerMutationType extends BaseMutationType
 	public static function mutationType(): array
 	{
 		return [
-			'type' => Types::partnerMutation(),
+			'type' => MutationTypes::partnerMutation(),
 			'args' => [
 				'id' => Type::int(),
 			],
@@ -94,6 +81,29 @@ final class PartnerMutationType extends BaseMutationType
 				'type' => Type::int(),
 				'description' => 'Идентификатор категории',
 			],
+		];
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function getConfig(): array
+	{
+		return [
+			'fields' => [
+				'update' => [
+					'type' => ErrorTypes::validationErrorsUnionType(QueryTypes::partner()),
+					'description' => 'Обновление партнера',
+					'args' => $this->getArgs(),
+					'resolve' => fn(array $rootArgs, array $args = []): array => $this->update($rootArgs, $args),
+				],
+				'create' => [
+					'type' => ErrorTypes::validationErrorsUnionType(QueryTypes::partner()),
+					'description' => 'Создание партнера',
+					'args' => $this->getArgs(),
+					'resolve' => fn(array $rootArgs, array $args = []): array => $this->create($args),
+				],
+			]
 		];
 	}
 }
